+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 ~~~~~
 * Comparative timeline to the home page.
 * Zoom feature to the timeline grid.
+* Sorting options to the base settings.
 
 Changed
 ~~~~~~~